Extracting all the arguments submitted to generate an object

The function model.list returns the arguments submitted to the function used to generate the object of class "twoStepsBenchmark", "threeRuleSmooth" or "praislm".

Usage

model.list(object)

Arguments

object

an object of class "twoStepsBenchmark", "threeRuleSmooth" or "praislm".

Value

a list containing every evaluated arguments

Details

These are returned as they are after evaluation, model.list doesn't return a call.

Examples

benchmark <- twoStepsBenchmark(turnover,construction); model.list(benchmark)
#> $hfserie
#>            constant   hfserie
#> Jan 2000 0.08333333  52.95220
#> Feb 2000 0.08333333  52.23017
#> Mar 2000 0.08333333  53.90119
#> Apr 2000 0.08333333  54.46341
#> May 2000 0.08333333  55.18790
#> Jun 2000 0.08333333  55.47950
#> Jul 2000 0.08333333  57.00284
#> Aug 2000 0.08333333  56.85169
#> Sep 2000 0.08333333  57.33163
#> Oct 2000 0.08333333  57.80456
#> Nov 2000 0.08333333  58.05515
#> Dec 2000 0.08333333  59.70457
#> Jan 2001 0.08333333  59.60711
#> Feb 2001 0.08333333  60.82968
#> Mar 2001 0.08333333  60.79504
#> Apr 2001 0.08333333  60.64843
#> May 2001 0.08333333  61.22578
#> Jun 2001 0.08333333  61.33095
#> Jul 2001 0.08333333  60.32609
#> Aug 2001 0.08333333  62.70687
#> Sep 2001 0.08333333  61.04050
#> Oct 2001 0.08333333  63.25204
#> Nov 2001 0.08333333  63.03554
#> Dec 2001 0.08333333  63.68417
#> Jan 2002 0.08333333  62.95983
#> Feb 2002 0.08333333  62.74072
#> Mar 2002 0.08333333  62.93436
#> Apr 2002 0.08333333  64.92714
#> May 2002 0.08333333  64.63393
#> Jun 2002 0.08333333  64.38931
#> Jul 2002 0.08333333  64.31761
#> Aug 2002 0.08333333  65.38775
#> Sep 2002 0.08333333  65.54656
#> Oct 2002 0.08333333  66.19863
#> Nov 2002 0.08333333  64.79341
#> Dec 2002 0.08333333  65.32345
#> Jan 2003 0.08333333  65.91953
#> Feb 2003 0.08333333  66.96147
#> Mar 2003 0.08333333  66.95276
#> Apr 2003 0.08333333  67.02189
#> May 2003 0.08333333  65.94428
#> Jun 2003 0.08333333  67.71240
#> Jul 2003 0.08333333  67.92858
#> Aug 2003 0.08333333  69.34735
#> Sep 2003 0.08333333  68.82502
#> Oct 2003 0.08333333  69.64556
#> Nov 2003 0.08333333  69.56173
#> Dec 2003 0.08333333  70.20201
#> Jan 2004 0.08333333  71.62133
#> Feb 2004 0.08333333  71.60853
#> Mar 2004 0.08333333  71.43942
#> Apr 2004 0.08333333  72.86131
#> May 2004 0.08333333  72.63120
#> Jun 2004 0.08333333  72.66541
#> Jul 2004 0.08333333  73.45298
#> Aug 2004 0.08333333  72.14682
#> Sep 2004 0.08333333  74.14137
#> Oct 2004 0.08333333  74.75602
#> Nov 2004 0.08333333  75.01225
#> Dec 2004 0.08333333  75.16693
#> Jan 2005 0.08333333  74.78895
#> Feb 2005 0.08333333  75.03312
#> Mar 2005 0.08333333  75.46066
#> Apr 2005 0.08333333  76.25605
#> May 2005 0.08333333  76.12112
#> Jun 2005 0.08333333  76.67231
#> Jul 2005 0.08333333  78.49390
#> Aug 2005 0.08333333  78.99697
#> Sep 2005 0.08333333  79.85180
#> Oct 2005 0.08333333  79.72473
#> Nov 2005 0.08333333  81.93822
#> Dec 2005 0.08333333  80.16945
#> Jan 2006 0.08333333  81.32016
#> Feb 2006 0.08333333  82.35133
#> Mar 2006 0.08333333  83.63386
#> Apr 2006 0.08333333  83.63608
#> May 2006 0.08333333  85.69523
#> Jun 2006 0.08333333  86.73240
#> Jul 2006 0.08333333  87.57389
#> Aug 2006 0.08333333  87.21008
#> Sep 2006 0.08333333  88.40514
#> Oct 2006 0.08333333  88.04695
#> Nov 2006 0.08333333  89.63659
#> Dec 2006 0.08333333  91.92662
#> Jan 2007 0.08333333  92.63169
#> Feb 2007 0.08333333  93.73279
#> Mar 2007 0.08333333  94.27983
#> Apr 2007 0.08333333  93.66900
#> May 2007 0.08333333  96.34922
#> Jun 2007 0.08333333  96.29693
#> Jul 2007 0.08333333  96.11316
#> Aug 2007 0.08333333  96.19652
#> Sep 2007 0.08333333  96.89635
#> Oct 2007 0.08333333  97.79498
#> Nov 2007 0.08333333  96.19817
#> Dec 2007 0.08333333  98.43091
#> Jan 2008 0.08333333 101.61886
#> Feb 2008 0.08333333 102.38942
#> Mar 2008 0.08333333 105.01997
#> Apr 2008 0.08333333 102.78853
#> May 2008 0.08333333 101.60734
#> Jun 2008 0.08333333 103.35920
#> Jul 2008 0.08333333 103.23944
#> Aug 2008 0.08333333 104.01456
#> Sep 2008 0.08333333 102.98514
#> Oct 2008 0.08333333 103.56430
#> Nov 2008 0.08333333 100.20945
#> Dec 2008 0.08333333 100.93351
#> Jan 2009 0.08333333  99.10659
#> Feb 2009 0.08333333  98.94825
#> Mar 2009 0.08333333  98.60390
#> Apr 2009 0.08333333  98.76337
#> May 2009 0.08333333  97.40919
#> Jun 2009 0.08333333  97.63918
#> Jul 2009 0.08333333  98.02665
#> Aug 2009 0.08333333  96.94162
#> Sep 2009 0.08333333  96.01264
#> Oct 2009 0.08333333  96.65990
#> Nov 2009 0.08333333  96.21037
#> Dec 2009 0.08333333  97.86338
#> Jan 2010 0.08333333  95.96627
#> Feb 2010 0.08333333  94.16255
#> Mar 2010 0.08333333  93.94231
#> Apr 2010 0.08333333  94.30199
#> May 2010 0.08333333  95.66035
#> Jun 2010 0.08333333  96.64727
#> Jul 2010 0.08333333  97.35644
#> Aug 2010 0.08333333  96.81579
#> Sep 2010 0.08333333  98.92692
#> Oct 2010 0.08333333  98.26915
#> Nov 2010 0.08333333 100.43624
#> Dec 2010 0.08333333  99.46794
#> Jan 2011 0.08333333 100.06816
#> Feb 2011 0.08333333 100.29381
#> Mar 2011 0.08333333 101.21072
#> Apr 2011 0.08333333 101.59274
#> May 2011 0.08333333 102.60177
#> Jun 2011 0.08333333 101.83867
#> Jul 2011 0.08333333 104.77932
#> Aug 2011 0.08333333 101.99032
#> Sep 2011 0.08333333 102.77701
#> Oct 2011 0.08333333 102.74454
#> Nov 2011 0.08333333 105.62406
#> Dec 2011 0.08333333 105.89324
#> Jan 2012 0.08333333 104.25699
#> Feb 2012 0.08333333 102.88971
#> Mar 2012 0.08333333 104.90658
#> Apr 2012 0.08333333 104.73297
#> May 2012 0.08333333 105.38963
#> Jun 2012 0.08333333 104.45144
#> Jul 2012 0.08333333 106.44087
#> Aug 2012 0.08333333 104.55779
#> Sep 2012 0.08333333 103.91388
#> Oct 2012 0.08333333 104.87647
#> Nov 2012 0.08333333 103.07460
#> Dec 2012 0.08333333 103.09632
#> Jan 2013 0.08333333 104.99670
#> Feb 2013 0.08333333 103.41157
#> Mar 2013 0.08333333 100.96894
#> Apr 2013 0.08333333 104.85869
#> May 2013 0.08333333 104.69964
#> Jun 2013 0.08333333 103.81736
#> Jul 2013 0.08333333 105.36523
#> Aug 2013 0.08333333 105.82866
#> Sep 2013 0.08333333 104.66512
#> Oct 2013 0.08333333 103.93619
#> Nov 2013 0.08333333 106.27217
#> Dec 2013 0.08333333 107.68245
#> Jan 2014 0.08333333 101.52057
#> Feb 2014 0.08333333 103.54678
#> Mar 2014 0.08333333 103.19782
#> Apr 2014 0.08333333 102.40583
#> May 2014 0.08333333 100.65119
#> Jun 2014 0.08333333 103.75838
#> Jul 2014 0.08333333 101.13113
#> Aug 2014 0.08333333 104.51738
#> Sep 2014 0.08333333 101.90654
#> Oct 2014 0.08333333 101.48923
#> Nov 2014 0.08333333  99.86480
#> Dec 2014 0.08333333 101.33127
#> Jan 2015 0.08333333 100.13741
#> Feb 2015 0.08333333 100.08938
#> Mar 2015 0.08333333 100.17402
#> Apr 2015 0.08333333  99.86940
#> May 2015 0.08333333  98.00527
#> Jun 2015 0.08333333 100.67095
#> Jul 2015 0.08333333 100.16710
#> Aug 2015 0.08333333  99.65097
#> Sep 2015 0.08333333  99.56027
#> Oct 2015 0.08333333 100.34261
#> Nov 2015 0.08333333 101.35284
#> Dec 2015 0.08333333  99.97979
#> Jan 2016 0.08333333  99.31158
#> Feb 2016 0.08333333 100.97860
#> Mar 2016 0.08333333  99.01797
#> Apr 2016 0.08333333  98.35013
#> May 2016 0.08333333 100.73473
#> Jun 2016 0.08333333  98.38691
#> Jul 2016 0.08333333  98.68740
#> Aug 2016 0.08333333  98.90801
#> Sep 2016 0.08333333 101.66512
#> Oct 2016 0.08333333 101.01954
#> Nov 2016 0.08333333 102.88054
#> Dec 2016 0.08333333 101.81342
#> Jan 2017 0.08333333 102.81422
#> Feb 2017 0.08333333 104.23883
#> Mar 2017 0.08333333 107.09768
#> Apr 2017 0.08333333 102.40725
#> May 2017 0.08333333 109.29157
#> Jun 2017 0.08333333 111.01618
#> Jul 2017 0.08333333 108.47322
#> Aug 2017 0.08333333 106.92717
#> Sep 2017 0.08333333 108.17751
#> Oct 2017 0.08333333 108.29035
#> Nov 2017 0.08333333 109.25391
#> Dec 2017 0.08333333 112.38608
#> Jan 2018 0.08333333 111.59429
#> Feb 2018 0.08333333 111.82158
#> Mar 2018 0.08333333 107.70471
#> Apr 2018 0.08333333 112.34397
#> May 2018 0.08333333 109.13911
#> Jun 2018 0.08333333 113.29352
#> Jul 2018 0.08333333 111.02063
#> Aug 2018 0.08333333 114.74015
#> Sep 2018 0.08333333 112.98397
#> Oct 2018 0.08333333 116.52727
#> Nov 2018 0.08333333 115.37020
#> Dec 2018 0.08333333 117.24465
#> Jan 2019 0.08333333 117.52315
#> Feb 2019 0.08333333 119.36003
#> Mar 2019 0.08333333 122.41181
#> Apr 2019 0.08333333 122.62517
#> May 2019 0.08333333 122.72672
#> Jun 2019 0.08333333 119.75711
#> Jul 2019 0.08333333 120.04214
#> Aug 2019 0.08333333 127.11451
#> Sep 2019 0.08333333 123.72686
#> Oct 2019 0.08333333 122.93404
#> Nov 2019 0.08333333 123.35912
#> Dec 2019 0.08333333 122.75935
#> Jan 2020 0.08333333 123.39553
#> Feb 2020 0.08333333 118.66687
#> Mar 2020 0.08333333 112.93565
#> Apr 2020 0.08333333  97.90891
#> May 2020 0.08333333  89.94552
#> 
#> $lfserie
#> Time Series:
#> Start = 2000 
#> End = 2019 
#> Frequency = 1 
#>  [1] 138.4 143.9 147.3 155.9 169.3 181.4 199.2 216.7 227.0 211.8 210.1 220.5
#> [13] 219.9 220.1 216.3 208.5 209.0 221.6 231.3 245.1
#> 
#> $include.rho
#> [1] FALSE
#> 
#> $include.differenciation
#> [1] FALSE
#> 
#> $set.coefficients
#> numeric(0)
#> 
#> $start.coeff.calc
#> NULL
#> 
#> $end.coeff.calc
#> NULL
#> 
#> $start.benchmark
#> NULL
#> 
#> $end.benchmark
#> NULL
#> 
#> $start.domain
#> NULL
#> 
#> $end.domain
#> NULL
#>